starting to refactor commands
